+++ tests/testthat/test-quarto-helpers.R
@@ -1,7 +1,7 @@
 test_that("Quarto Outputs", {
   out = list()
   out$posit_version = "2022.10.0"
-  msg = get_quarto_posit_version_msg(out, "connect")
+  msg = get_quarto_posit_version_msg(out, type = "connect")
   expect_true(stringr::str_detect(msg, "CVEs"))
 
   out$posit_version = "2012.10.0"